Remove PACKMSG_NEGATIVE_FIXINT.
We have PACKMSG_POSITIVE_FIXINT, because of the ambiguity that it can be
read by both packmsg_get_int*() and packmsg_get_uint*() functions. However,
there is no need for PACKMSG_NEGATIVE_FIXINT, it can only be read by
packmsg_get_int*() functions, so we can just as well report it as
PACKMSG_INT8.